+++ src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Route from 'react-router-dom/Route';
 import Switch from 'react-router-dom/Switch';
-import Home from './app/components/Home';
+import Home from './app/components/Home.jsx';
 import './styles/App.scss';
 
 const App = () => (
