+++ src/styles/Home.scss
@@ -1,5 +1,5 @@
 body {
-  background-color: #ffebf5;
+  background-color: #58a5f0;
 }
 
 .Home {
@@ -12,15 +12,19 @@ body {
 }
 
 .Home-header {
-  background-color: #ffd8eb;
-  //background-color: #222;
-  height: 150px;
+  background-color: #0277bd;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  height: 250px;
   padding: 20px;
   color: white;
 }
 
 .Home-intro {
   font-size: large;
+  padding: 1rem;
 }
 
 .Home-resources {
