+++ stories/index.stories.js
@@ -4,6 +4,7 @@ import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import { linkTo } from '@storybook/addon-links';
 import Home from '../src/app/components/Home';
+import Technologies from '../src/app/components/Technologies';
 
 import { Button, Welcome } from '@storybook/react/demo';
 
@@ -19,5 +20,6 @@ storiesOf('Button', module)
     </Button>
   ));
  
-storiesOf('Plop', module)
-  .add('Home', () => <Home />);
+storiesOf('Site', module)
+  .add('Home', () => <Home />)
+  .add('Technologies', () => <Technologies />);
